pyproject.toml
[project]
name = "agentcore-efs"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "bedrock-agentcore-starter-toolkit",
]
